test: fold Anthropic handler suites
